divide logic for importer to dedicated files
